Transient properties
Added support for transient properties. Right now the support for these
properties still stores physically. This is done that way because if we'd use
TEMPORARY tables in sqlite, the table wouldn't be shared between the processes
of trackerd and tracker-indexer. When we refactor to letting trackerd do both
writes and reads, we can instead start using TEMPORARY tables. I have added
the code for this, but commented out in this commit.