Disables evolution push module and improves the code
This patch also fixes the coding style in places and replaces some
strchr() calls for the utf8 equivalent. Code was added to make the
evolution plugin work. The "File:Modified" metadata was missing and
that was the reason nothing was being added to the database.

NOTE: We are not using this because doing so makes the indexer
completely unresponsive. This is what is needed to actually make the
evolution plugin feed our database.  However, there is no queueing, no
pausing, no control at all, we simply handle everything as it comes.
This is not good enough for tracker-0.6 because it causes a regression
in the indexer. So this is disabled.

NOTE: This code is supposed to be quite different in the tracker-0.7
branch and as such it makes no sense fixing it.