Fixes: GB#572237, Remove deprecated Glib symbols
This reworks the find_orig_uri() which was broken. This now uses
g_utf8_casefold() instead of g_utf8_up(). The function now oly uses
g_utf8_casefold() as a last resort after checking the extensions and
file "name" parts found in the directory do not match. This should be
more efficient since string creation only occurs as a last step. Also
the info and orig_info were not being compared in the first place
which is another bug fixed here.