Fixes: NB#114354, trackerd polls
This was more work than I expected. I think I fixed another bug with
the testing on this one. We now make sure that if we go back to IDLE
state after we have finished PENDING/WATCHING states for crawling we
correct the state to go straight back to PAUSED if we are paused for
any of the reasons we can be paused. The bug here was that we ended up
indexing content if we paused based on a full disk if the full disk
state occurred BEFORE we finished crawling the file system.

Also, we now watch the TrackerConfig low-disk-space-limit notify
signal to update in real time.