Get the actual mount point for a given path to get volume id in db
Before we did this, we were using clever SQL statements to get the
ID in the Volumes table based on a given path and the known volume's
mount point in the database. This of course doesn't give a real idea
of the mount point because you can't rely on clever string
manipulation, only the file system really knows which removable device
the file is on. So now we get the real device and use some extra
caching to improve performance here too so we can get the UDI by path.
The UDI is then looked up in the database for the volume ID.

NOTE: This is related to and improves on NB#109892 and NB#112632.