Set up volume states before enabling public APIs via DBus
In the startup for the daemon, we now make sure we free
TrackerMainPrivate and fix a warning there. We also change the startup
so we wait until the indexer has replied regarding setting up volumes
before we allow users to send DBus requests in. If the indexer fails
to reply successfully with the volumes set up, we will exit
immediately. The reason for this, is because the SQL queries highly
depend on the Volumes table being accurate. A restart (triggered from
another DBus request) may well fix this anyway.

In the daemon module, we don't get the stats when we create the
GObject because tracker-main requests that for us AFTER we have set up
all our public interfaces (DBus, push modules, etc).

In the processor, we now set ->finished to FALSE when starting and we
also check ->finished before in _stop() so we don't print processor
summaries when we have already finished.