Pause indexer when setting volume states & disable TTL use
Disable udi variable check in tracker-data-update, was causing some
log warnings when NULL is quite acceptable.

Disabled TTL APIs (tracker_removable_device_*) in indexer which was
causing several bugs related to speed and caching of albumart.

Make sure we quit the main loop in the event we have NO volumes to
enable on startup in the daemon.

Make sure we pause the indexer before doing ANY volume calls to ensure
they get through when the indexer is quite busy. Any failure response
here will shutdown the daemon so this is quite important.