Introduce TrackerStore, a queue for updating
TrackerStoreQueue is or will be the internal API for most data access and
storage. This implementation uses the GMainLoop and a GQueue for all batch
requests.

Due to its use of the GMainLoop the API is going to block the mainloop, but it
wont block the caller's code. It will instead find a nice slot during the
GMainLoop the execute itself on. When finished it'll execute first your
callback and then your GDestroyNotify. All parameters that you receive in the
callback that aren't user_data must not be freed (they are to be considered
read-only in your callback).