Monitor cache time now 1s & processor queue time now 0s for < 50 items
Now we only wait 1 second before sending files to the processor from
the monitor module. The reason for this, is to be more responsive from
monitor events. Now a DELETE event is handled within 1 to 2 seconds if
the indexer is idle.

The processor queue handler is set up using a 2 second timeout per
1000 files. Now when we have finished the initial check of all files,
we use g_idle_add() instead to make sure the indexer deals with items
much quicker.