The solution to the 2.5s per 10k items: simply lower the priority of the queue
This is my analysis about this: It was actually clear when Jürg told me did
his tests without letting the store-queue's idle handler run. My first idea was
to also disable it for my tests, and then suddenly 10k inserts finished within
0.135s instead of 2.5s (latter was a quite dissapointing result, former is the
expected result).

My analysis is the following: the queue's GSource had an equal priority as the
unix socket's GSource of the GIOChannel.

Meaning that if we lower the priority of the queue's GSource, lower than the
GSource of the unix socket's GIOChannel, that the latter will run at speed.

With this mini-patch test tests/estress/socket_stress_test.c will finish in
about 0.135s on an average desktop. The queue will of course have a lot of
work afterwards, but any subsequent request on the unix-socket (and probably
also on DBus, as also DBus's GSource now runs at a higher priority that the
queue's) will get priority.

This allows us to queue 10000 INSERT queries in 0.135s instead of 2.5s and yet
we don't need to halt or disable the queue itself.

I made this rather large analysis because this apparently is what makes the
difference in IPC responsiveness for tracker-store's queue design.