Fixed state machine so we return to previous state after PAUSED
This bug was simulated by setting the LowDiskSpaceLimit in the config
to 99 and then starting the daemon. Then setting it to 1 in real time
should start sending files to the indexer, but instead it was returned
to PENDING state which meant the queue handler returned and didn't do
anything since it was expecting IDLE. IDLE was attempted but denied
given pause conditions.