Fixes: NB#119230, tracker-extract: memory management shall be improved
This is another stab at using setrlimit(). For now we set the memory
limit to 80Mb as default and 512Mb for AMD. For some reason, anything
less than ~450Mb doesn't even allow tracker-extract to dlopen the
modules when we use GStreamer.

Doing a quick test with tracker-stats before and after shows it is
working on my AMD64 desktop and still indexing the same amount of
content.