Fix not crawling inserted removable media
Now we start the processor again when we get new removable media
inserted, this bug mainly affects the condition when we have finished
processing all files and the removable media state changes again.

The statistics we report about files crawled and time taken is now
reset and correct each time this occurs.

The mount point added and removed callbacks from HAL were removed in
TrackerProcessor and now two functions are called FROM tracker-main.c
which do the same thing. The reason for this, is that we change state
while setting volume states and we can't process removable media in
some states. So to control this we do everything AFTER we have set the
removable media states correctly.