Fix NULL crasher in trackerd with strcasecmp
This is down to not checking variables in libtracker-data before we
start using them and by passing dodgy variables to tracker-metadata in
the daemon. Now we have a check in the daemon AND variable checks in
libtracker-data too.

Also, I made the ASC and A comparison in the daemon case insensitive.