tracker-extract: push all albumart code to tracker-miner-fs
Now we signal data that is used to be processed for albumart. This is
all handled in the miner-fs now, so the extractors simply say, here is
my data and length of data (and some other details like artist), now
process it. The miner-fs then uses the code previously
tracker-extract-albumart.c to either download the art, use GdkPixbuf
or save the data directly to the right filename. The thumbnails are
then requested staight after.