Fixes NB#129415, Checked] [REGF] Video and Audio Library view and ...
This was caused by Microsoft Windows NOT setting the parent folder
mtime when it changes a child file. On Linux we can count on this. On
Windows we can not. So now we check in ALL cases for EVERY folder if
we have any stale children in the database and we remove them if
that's the case.