Add tracker-miner-discover to find DBus miner services
This also makes use of tracker-crawler to find .service files in the
DBus services directory, the debugging from tracker-crawler was
unavoidable so the g_debug and g_message statements have been moved to
tracker-miner-process instead so it can decide if it wants to print
found/ignored things.

Also the tracker-crawler finished signal now includes WAS_INTERRUPTED
boolean so we know if it was stopped prematurely.