Fixes: NB#138860, tracker-miner-fs stops crawling if a directory...
This was caused by returning TRUE to silently handle missing
directories, the problem is, we then depend on a "finished" signal
from the crawler which we never get. As such, we now return FALSE in
the _start() function if we don't expect to progress any further.

Also, we now don't reset the GCancellable and don't set the crawler
into state "running" until the initial checks are done first.