tracker-miner-fs: Use XDG spec for special user dirs in config
This means now the user can specify "&USER_DIRECTORY_DOCUMENTS" and it
will point to whatever the XDG config says in
/etc/xdg/user-dirs.defaults and use those dynamically.

For now saving the config back is not supported so well. This can be
improved in the future and it is currently unused in tracker-miner-fs
anyway. The preferences need some love after this too.