All miner SPARQL Update queries need FROM or INTO to avoid writeback
When INTO or FROM are passed, graph in tracker_writeback_check wont be NULL.
A NULL graph means 'default graph' or 'something that came from the user'.
We only want to do writeback for things that came from the user. This is how
we distinguish between content from the miner and content from the user.