Handling writeback in tracker-miner-fs
When a writeback takes place, tracker-miner-fs must NOT create a SPARQL Update
that contains any predicates that are marked as tracker:notify. For this
reason I added a mechanism for marking a subject in tracker-miner-fs's queues:
The first next time the subject is to be analyzed it checks whether the
subject was marked. If it was marked then instead of doing a full extraction,
tracker-miner-fs will only get a few properties (like mime, mtime, size).

* MIME because the content-type of a file might have changed by writeback
* Size because the size of the file might have changed by writeback
* Mtime because the writeback will update the file's mtime, and we want
  to sync tracker-store with that too