Autogenerate ontology documentation.
There is now a new gtk-doc reference for ontology, which is created
from data/ontologies/*. docs/tools contains the programs to generate
both docbook and graphs, these are dumped on docs/reference/ontology,
which is compiled afterwards.

Documentation generation happens on make, ontology graphs require
graphviz and are only generated on distcheck, A base image is
committed, so people don't need graphviz to have a peek.