src/plugins/nautilus: Fixed crasher when right clicking on background
This was spotted by Debarshi Ray.

There is no action for the background now, unless we decide to tag the
background at some point, this won't be enabled (for now disabled).

If this code is enabled it won't crash, however, we can only list
tags, we have no file to apply them to.