libtracker-miner: Fixed evolution plugin enable/disable calling exit()
This occurred because we didn't release the name when we clean up the
dbus data for the miner. Subsequently, disabling and then enabling the
plugin would try to re-register the miner's name and find it was in
use and so exit() would be called in libtracker-miner. Now this
is fixed and Evolution doesn't completely exit.