TrackerMonitor: Workaround design flaw in libinotify.
libinotify keeps a hashtable of watch descriptor/handle, and does
removing in an asyncrhonous manner. So when inotify_add_watch()
reuses a watch descriptor, removing/adding whithout having
libinotify process the events coming from the fd may cause the
wrong handle (the just added one) to be removed.

TrackerMonitor now adds, and then removes old monitors, when a
directory is moved, so there's no opportunity to remove the wrong
handle.