Fix G_IS_OBJECT(file) errors.
the ProcessData was being found through g_file_equals(), but
under a fast sequence of create/remove/create events for a file
(as autofoo likes to do), there could be several ProcessData
for different GFiles pointing to the same path. So
process_data_find() now checks for pointer equality, since the
GFile tracker_miner_fs_notify_file() is being called on should
correspond to the GFile passed when emitting ::process-file.