TrackerCrawler: Report a GNode with the file hierarchy.
TrackerCrawler now provides a GNode with the directory/file
hierarchy on ::directory-crawled for each GFile passed at
tracker_crawler_start(), all the stats are also reported
per directory root. ::finished is now only used to notify
about interruption.

TrackerMinerFS now takes a copy of each of the GNodes intead
of appending directly to items_created. Whenever items_created
is empty, the stored GNodes are checked for more contents to
inspect. This is done so TrackerMinerFS guarantees that a
directory has been already added to the store before processing
its children.