TrackerMinerFS: Add API to know a file/parent urn.
This API allows retrieving the URN for a GFile being
processed (to check for file existence in the store),
or it's parent's (to set nfo:belongsToContainer
relationship).

Since files are guaranteed to be processed together
with the other files in the same folder, and parent
folders are guaranteed to be fully processed before
its children, The method to cache the parent folder
URN should have little impact on performance.

In order to retrieve the URN for the file itself,
a extra call to item_query_exists() had to be added,
unfortunately, doing so on should_change_index_for_file()
is not an option, since it happens on the crawling
phase, so all URNs would have to be stored. This is
a performance vs memory usage problem.