Bug #609004 - Timeout on 1 file during extraction causes 10 to fail .
TrackerMinerFS can process paralelly as many files as the ::process-pool-limit
property say, but tracker-extract inspects files one by one. So if some file
takes longer than the extractor timeout, all the other concurrently processed
files will receive a DBus no reply/timeout as well.

So, if we get a DBus timeout on any file that's not the first that was added
to the processing pool, we retry extraction on that file.