Fixes NB#158450: Tracker doesn't remove song from database when ...
Remove too lax check for not monitoring IndexSingleDirectories children, which
was also affecting legit dirs in IndexRecursiveDirectories which were a child
of the former.

TrackerMinerFS should only emit monitor_directory() for directories where both
check_directory() and check_directory_contents() have returned TRUE, We're
actually interested in monitoring any directory (and children) in
IndexRecursiveDirectories where that does happen, and the latter call shouldn't
happen for children of IndexSingleDirectories, so it's safe to just return TRUE
there.