libtracker-fts: Separate FTS index from virtual table initialization
As we do not use the virtual table layer for updates, the combined
initialization of FTS index and virtual table layer causes problems
when SQLite decides to disconnect the virtual table. This happens
on database schema changes.

With this patch the FTS index can still be used even if the virtual
table layer is disconnected.