tracker-preferences: Workaround for gtk+ bug setting scalers to max default
The scaler type in gtk+ seems to be setting the scaler's value to its max
default when the scaler is in a tab, the tab isn't open and
gtk_widget_show_all is called for. It also signals when this happens the
changed signal. In the signal handler we set the config object. If the tab
is never activated that means that we've written the config value with some-
thing the user didn't ask for (the default max value). I workaround-fixed
this by only setting the config object on apply-click.