libtracker-miner: Add TrackerNetworkProvider class
The TrackerNetworkProvider class provides functions to get information
about the network status, such as if there is an active connection or
not, and what the connection type is (GPRS, 3G, WIFI etc.). It can be
used by web miners using web services to avoid heavy transfers when
being connected on a slow and/or expensive connection.