libtracker-data, libtracker-db: Reimplement Backup() and Restore() to use tar
The original implementation of Backup() and Restore() used a simple file copy
of the journal. When we have rotated journals we of course can't use only the
active journal for Backup(). And for Restore() we want to put the rotation state
back the way it was when the backup was made too. So the best option I could
come up with was to use a simple tar.