Add Tracker Steroids interface
The Steroids interface uses a local pipe to transfer query results,
instead of sending them over DBus. To do so, it uses a new feature in
DBus 1.3 which allows passing UNIX file descriptors in messages.

This commit introduces a new DBus object exported on the store at path
/org/freedesktop/Tracker1/Steroids. Methods are in the interface
org.freedesktop.Tracker1.Steroids.

The DBus interface is composed of two methods:
PrepareQuery <- SPARQL query
             -> UNIX file descriptors where results will be written
             -> Query identifier
The PrepareQuery call initiates a query in Tracker. The query is not
really prepared, in the sense that a PrepareQuery with an erroneous
query will succeed. It's the Fetch call which will fail in that case.
The returned query identifier can be used to initiate the Fetch.

Fetch        <- Query identifier
The Fetch call executes the query with the given identifier, and returns
the results over the pipe returned by PrepareQuery.

Note that while results are being fetched, a DB cursor is left open,
blocking other concurrent INSERTs. It is therefore important to fetch
the results and free the iterator as fast as possible.