libtracker-client: Add Steroids support
This commit add several function in libtracker-client to exploit the new
Steroids interface.

The main function is tracker_resources_sparql_query_iterate, which will
return a TrackerResultIterator. This iterator can then be used to
iterate over the results using the tracker_result_iterator_* functions.
Note that iteration is not done on DB side, all results are first fetch
into a buffer on client side. This is because keeping an iterator on
server side would block access to other clients, SQLite not being MVCC.