--enable-maemo: had no effect
enable_maemo was never set to anything but "no" in the default case in
configure.ac.

Note that enable_maemo=yes in the "enable option selected" case would
be wrong, because --disable-maemo is valid and ends up in that branch
with $enableval == "no". Some other uses of AC_ENABLE() make that mistake
(not fixed in this patch).