libtracker-db: Add a "length" parameter to tracker_db_cursor_string
This commit adds a "length" parameter to tracker_db_cursor_string, so
that the client can know the length of the string, without needing to
call strlen. The function internally used to obtain the length can be
faster than strlen.

This commit also port existing components to the new prototype.