Fixes NB#174333 - Each file is listed twice in tracker search results
Unset parent folder so caches are regenerated.

The issue was due to priv->current_folder being reused for both mtime
cache (for crawling stage) and iri cache (for indexing stage). So when
recrawling, the miner changes current_folder without invalidating the
iri cache, so there could be a mismatch when it enters the indexing
stage again.