NFO: After careful measuring we noticed that this index isn't faster
The index for nfo:belongsToContainer was not faster while testing with
the Linux source code directory. We did several runs to avoid hot vs.
cold cache measurement differences. We also tried removing the index
on nfo:fileLastModified but this didn't result in a significant
difference (so we're leaving that index for now).