Fixes: GB#622702 - 0.8.13 does not build with gio 2.22 anymore
Bumped version of GLib required from 2.20 to 2.24

This is due to commit d0f60bf212624de4aa41a01ab6d5cd7257e5ea50
Unfortunately, this could not be reverted because other code has been
applied on top of it. For now, the easiest thing to do is to bump the
GLib version. Version 2.24 should be quite available now.