tracker-store, libtracker-data: Workaround for graph not being used in SPARQL DELETE
Because GRAPH is kept NULL for tracker-data-update.c during a SPARQL DELETE with a
graph wasn't the writeback functionality being done correctly: only for the
default graph must the writeback signal emit. But with graph always being NULL for
deletes, would it always emit for any delete.

And the IgnoreNextUpdate SPARQL query of tracker-miner-fs does a DELETE in a graph,
meaning that this triggers tracker-store into doing a new writeback emit, and makes
tracker-writeback writeback again, and that makes miner-fs do a new IgnoreNextUpdate,
until the end of times (endless loop between store, writeback and miner-fs)