tracker-store: Ensure inline compilation of alloca by adding the include
Check man page of alloca:

Normally, gcc(1) translates calls to alloca() with inlined code. This is
not done when either the -ansi, -std=c89, -std=c99, or the -fno-builtin
option is given (and the header <alloca.h> is not included). But beware!
By default the glibc version of <stdlib.h> includes <alloca.h> and that
contains the line:

So just to be sure ... :-)