libtracker-data: Fix for tracker:domainIndex ontology changing
When you set the tracker:domainIndex to a specific property twice,
for example two classes get nie:title, then the ontology change
coping was failing because the registration that the change has to
happen was done on nie:title. Making each occurrence of a domain-
index on nie:title be detected as changed. This commit fixes that.