tracker-writeback, libtracker-miner: Don't writeback non-strict mined values
The idea here is that since the 'guarantee' branch by Mikael that the
miner might 'guess' non-strict values heuristically. This isn't good
when the writeback service is enabled, as that would mean that the
writeback service could writeback heuristicly guessed values.

To fix this we require that the GRAPH of a value that is to be written
back isn't set to the miner-fs's GRAPH.