libtracker-data: After journal replay modseq must get restored
This commit changes how the modification-sequence works, this is
made public using the tracker:modified property on resource.

Instead of each resource having its own unique modification-
sequence number, the modification sequences are defined per
committed transaction that had properties that aren't transient.

This commit also changes the fact that a transaction with only
transient properties was still adding an empty transaction to
the journal. With that the modification sequence wouldn't
restore correctly to original state. It wasn't needed, so this
change removes the empty transaction in the journal in that
special case.