functional-tests: better logging. Reduce race-condition killing process
We are killing the process and then wait for the signal that it disappeared from dbus.
Sometimes the process dies very quickly and the name changed already happend,
so we only continue after a timeout.

Using a flag to store when the name changes to avoid waiting for the signal if we know
already the daemon is not there.