tracker-writeback: Add nfo:FileDataObject to tracker-writeback query
This blocks tracker-writeback from trying to writeback to so-called
virtual files (resources that are being filled in for a file that
will arrive as a file in nearby future).

It does this by requiring nfo:FileDataObject: so-called virtual files
must not be made nfo:FileDataObject but only nfo:DataObject, as it'll
be tracker-miner-fs picking up and filling in the nfo:FileDataObject
parts. As soon as this has happened, tracker-writeback is permitted
to start becoming responsible for writing metadata back into the file.

Fixes NB#222494.

However, this bug illustrates other problems that must be fixed too.