libtracker-sparql: Use separate thread for sync init if necessary
tracker_locale_init must run with the thread-default context set to the
global default context. As pushing a new thread-default context
requires the current thread to be the owner of the passed context, we
cannot always push the global default context as a thread-default
context.

This commit uses a threadpool job in the case that sync initialization
is requestesd with the thread-default context not set to the global
default context.