libtracker-common: Re-instate setrlimit()
Previously it was disabled due to GStreamer constantly causing high memory use.

Now we have re-enabled it and instead of a flat 512Mb (or 80Mb for 64bit) limit,
we use 50% of memory capping at MAXLONG (2Gb on 32bit machines) and using
at least 256Mb as a minimum.

This was done to fix my desktop with 4Gb going OOM from a broken PNG file