tracker-extract: Implement thread awareness for modules
Tracker extract modules may provide now an init() function, which
also reports the thread awareness of the given module, allowing
these to run sequentially in the main thread, on a dedicated thread
for the extract module, or in a thread pool.

At the moment no extractor modules do this, so they run by default
in the main thread as before.

Conflicts:

	src/tracker-extract/tracker-extract.c