libtracker-miner: Avoid possible race condition when cancelling extraction
Operations for files that aren't sent to the extractor were instead finished
an idle function, which could lead to double frees if the task was cancelled
right in between (one resulting from cancellable::cancel, the other when
dispatching the idle function).

So, instead, don't do that in an idle at all, there is no reason why
we can't call tracker_miner_fs_file_notify() right away.