libtracker-miner: Fix inconsistent status with progress signals
Quite often, we see progress of 100% and still have status messages
saying Processing... which is not true. So we use an idle timeout to
help that situation. This happens because miners can use the
g_object_set() API with multiple properties and we were signalling the
update between each property being set.

This commit also enforces setting the status to 'Idle' or
'Initializing' when the progress is either 1.0 or 0.0

The output shown in tracker-control is now much cleaner.

The trace macro was also added to follow state/progress changes along
with signal emissions more easily.