tracker-monitor: translate CREATE(a)+MOVE(a->b)=UPDATE(b)
Fixes NB#251032. Tracker-writeback often creates a temporary hidden
file that, after modification, is moved onto the original file location,
This tricked TrackerMonitor so that the create and move operations there
were translated as a create operation, and create operations are
compressed with delete events into a noop. So if a quick delete came
after, TrackerMonitor was emitting no signal for a truly deleted file.

Instead, translate create+move as update, the miners handle creates
and updates equally, and these won't get compressed with delete
events, so those are still emitted.