libtracker-miner: Add TrackerSparqlBuffer
This object inherits from TrackerTaskPool, and takes care of
dispatching tasks containing SPARQL updates.

It is meant to replace the sparql buffers in TrackerProcessingPool,
and in fact inherits much code from there, the main difference being
that there is no "processing" queue for already packed up batches,
instead it relies on linear processing of updates.

That last queue was sheldomly used (if at all), as it requires a
choking tracker-store that can't insert stuff at the same pace
that miner-fs processes it, this situation looks quite unlikely,
and makes the code simpler.