libtracker-miner: Ensure prioritized tasks always have parents indexed first
We now check before item_add_or_update() for those files that have a parent
that should have been indexed, but actually isn't.

In that case, the file is inserted back to the queue, prepended by its
parent to ensure it is indexed, if the file meeting these conditions is
deep in the hierarchy, the parent file will be treated the same, so
the operation will unroll until an indexed/root directory is found.