libtracker-miner: Avoid triggering too often the ensure_mtime_cache() hack
The hack introduced in commit 22087297 is meant to cope with the "delete
folder and move another on top" situation, forcing reindex if the operation
was in the middle of being processed. This situation is most relevant for
monitor events, as deletes can take quite some time to happen recursively.

However, on unclean shutdowns, this condition also happens often. Even if
the directory is already indexed in the store, it is pushed for mtime
check into the queues, which looks exactly the same to ensure_mtime_check().
The final effect was that plenty of files were being reindexed for the sake
of it.