build: Bump tracker_api_version to 0.12 but keep major/miner as 0.11
Add AX_DOTS_TO_UNDERSCORES macro to allow independent
TRACKER_API_VERSION_UNDERSCORES value (which is used to convert "0.12" to
"0_12".

We need this because for 0.11.x releases, we want the files to be named *0.12*
which requires changing only tracker_api_version in configure. This breaks
Makefiles which build introspection files because they use
$TRACKER_MAJOR_VERSION_$TRACKER_MINER_VERSION and that's not updated.
Additionally, $TRACKER_API_VERSION can't be used because it contains a dot not
an underscore. The only way to fix this was to do it in configure.