tracker-extract: Improve checks for mimetype guessing
Previously mimetype had to be NULL, but over dbus, it can only be an empty
string. The check has been improved here.

Additionally, we were leaking a GSimpleAsyncResult.

There was also an issue with error conditions of the mimetype check where we
would return and the client would be left hanging because we didn't complete
the simple_async_result in that condition. That's now been fixed and any error
reported by getting the mimetype also reported back to the client.