ontology: Remove tracker:writeback property from nao:prefLabel
Fixes NB#278806, this property only applies to nao:Tag entities, not
on files, so whenever a file with XMP keywords was added, the insertion
of the keywords would trigger a writeback petition in the miner. This
would add a writeback task that would issue a costly query only to find
out it doesn't directly affect any nfo:FileDataObject.

If several files with these characteristics were created, the effect
would be multiplied as writeback petitions would be requested faster
than delivered, filling the writeback queue and hence blocking any
other processing.