tracker-miner-fs: Improve query performance for updating tracker:available on mounting
Previously we had an OPTIONAL and FILTER in the query. This translated into
quite a delay when updating even a small number of resources in the database
and subsequently affected the crawling speed upon re-mounting.

Given we don't *ever* expect there to be resources for the given
nie:dataSource which are in the wrong tracker:available state, we don't need
the OPTIONAL or FILTER there.

This fixes NB#272441