build: Don't require explicitly GTK+ 3.x for nautilus, it can cause crashes
This is because nautilus uses GTK+ 2.x on some systems and we require 3.x,
this causes crashes and stack overflows, etc which we want to avoid.

Fixes GB#651815 (possibly)

There have been a number of crash situations reported with nautilus and all
random and impossible to reproduce. I am considering this the fix for now
since I see no other reason.