tracker-extract-gstreamer: Don't give each track its own URL
Start time inside the larger container is specified now using a
new property nfo:containerStartTime, and the track is related to the
container nfo:FileDataObject in the standard way (nie:isStoredAs).

This avoids problems of having a number of nfo:FileDataObject
resources all pointing to the same file, such as duplication of
file metadata and the need for special code to update the nie:url
on moves/renames.