tracker-extract: Add 'postupdate' field
Normally in Tracker the nfo:FileDataObject resource representing a file
and the nie:InformationElement subclasses representing its contents are
one and the same.

If a file contains more than one logical resource, the extractor needs
to create these separately and link them to the file resource using
nie:isStoredAs - this cannot be done until the file resource has been
inserted in the store, because it is difficult to know how to refer to
the file object. The 'postupdate' field provides a way to do this.