tracker-miner-fs, applications: Keep track of MLocale-reported locale
When MeegoTouch libraries are available, we use the MLocale Translation
catalogue to get translated application names in the Applications Miner.

Until now, we assumed that the language locale in MLocale would be completely
equal to the one stored in the GConf keys, but this is no longer true; the
'/meegotouch/i18n/language' key may not hold the country code, while the
MLocale-reported language now it does (NB#276653). Therefore, it is better to
keep track of the MLocale-reported locale instead of the GConf-retrieved one.

Fixes NB#282957