tracker-writeback: Do proper error handling in the writeback program
The writeback_file_finished in the miner-fs requires that in case no rename()
happens at the end (which is the case in case the module's function returned
FALSE at the end of tracker_writeback_file_update_metadata), that an error is
returned for the DBus Method PerformWriteback.

I encountered a situation where xmp_files_open_new returned NULL in the module
function writeback_xmp_update_file_metadata of tracker-writeback-xmp.c, this
resulted in a FALSE return causing a unlink instead of rename of the temporary
file. This made miner-fs's handling for that fail.