libtracker-miner: Calculate remaining time purely on extraction time
During crawling, the extraction queue may be mostly empty, but adding
up time that makes the remaining time calculations to be quite off
when little items have been processed so far.

So, keep a timer only for extraction time, which can be used for more
accurate measurements, as only extraction is left when tracker-miner-fs
gets to report progress/remaining time.