libtracker-miner,crawler: Use a cancellable per crawled directory
Reusing the cancellable might bring in issues if tracker_crawler_start()
(which resets the cancellable) is called shortly after tracker_crawler_stop().
So pending async tasks that do check for g_cancellable_is_cancelled() might
run into the wrong state. Possibly fixes NB#287480.

The cancellable might be used to cancel the current operation, but then
reset and reused for the next one, while the previous operation could
still be on the way to be cancelled. Instead, create a new cancellable
to handle any following task.