libtracker-miner: Traverse children selectively in TrackerFileSystem
Now, returning FALSE in the TrackerFileSystemTraverseFunc given to
tracker_file_system_traverse() means the node will not be recursed,
but the traversal will still continue.

This deviates a bit from g_node_traverse, but makes more sense given
the common operations in a filesystem representation.

This is now used in TrackerMinerFS to stop traversal on deleted
folders.