libtracker-miner: rephrase code comment
It'd be indeed good to have tracker-miner-fs monitor directories
triggering a TRACKER_FILTER_PARENT_DIRECTORY filter in case that
situation changes (ie. the file triggering the filter disappears),
but given Tracker has never behaved like that and noone ever
complained, lower the comment severity a bit, it's not something
we should rush on.