libtracker-miner: Change signedness of a field that is used to calculate
The ABS() where this field was used in had a calculation that could
result in a negative value within the ABS's brackets. Because it's
unrealistic that we'd ever need the value of this field to be larger
than int, we decided to just change it to signed int.

Fixes NB#290165.