evolution: Fixed includes ... yet again
So I have this time put each include in separate sections. The logic was
getting too complicated and it was hard to maintain. This is why single
include files make so much sense.

I've checked each release of the libraries we import from and the files do
exist for each of the versions we support.

I removed the camel include because all the Evolution includes include that
already.

I also removed the include for older than 2.91 versions which didn't exist
there and was reported under the bug below.

Fixes GB#672415