libtracker-miner: Avoid processing item queue while it is blocked
item_add_or_update_cb() had a call to item_queue_handlers_set_up().
This was unneccessary as the queue will already be running, except in
one the following cases:
 -  the extraction or SPARQL task pools are full
 - the queue is blocked waiting for a specific file to process
 - the queue is empty

Previously, while waiting for a file to process we might pop and
reenqueue the next file in the queue several times, which is a waste of
effort and makes it difficult to detect if the miner is stuck in an
infinite loop.