build: don't build-depend on Dia
Currently, the gtk-doc generator code in Tracker has a build dependency
on the Dia application in order to generate some images for the
documentation; the dependency is not optimal, as Dia is not available
on all distributions and platforms.

This patch changes the build system to add a specific update-diagrams
make target to be ran from the ontology docs directory (which updates
the images), and adds the images themselves under version control.

Fixes GB#688287