Use fts4 external content tables on Tracker FTS
FTS support on Tracker is now implemented using external content
support available in sqlite >= 3.7.9 FTS4.

FTS tables created this way fetch data from a single table/view,
for this purpose an intermediate view has been created to
interface with the FTS table. As this view is mainly queried by
ID (both when populating the FTS contents, and when querying
throught the FTS table), queries are fast enough on it.

As FTS indirectly points to the data on the real tables where tracker
data is stored, strings themselves are stored only once in the database,
so there is no impact in database size when compared to the previous
custom FTS code. Performance had little changes too from testing.